Update: Move same codes from all MathParsers into \MathParser\AbstractMathParser
