+++ src/MathParser/ComplexMathParser.php
@@ -1,4 +1,6 @@
-<?php namespace MathParser;
+<?php 
+
+namespace MathParser;
 
 /*
 * @package     Parsing
@@ -13,16 +15,8 @@
 use MathParser\Parsing\Parser;
 use MathParser\Lexing\ComplexLexer;
 
-class ComplexMathParser
+class ComplexMathParser extends AbstractMathParser
 {
-    /** StdMathLexer $lexer instance of StdMathLexer used for tokenizing */
-    private $lexer;
-    /** Parser $parser instance of Parsed used for the actual parsing */
-    private $parser;
-    /** Token[] $tokens list of tokens generated by the Lexer */
-    private $tokens;
-    /** Node $tree AST generated by the parser */
-    private $tree;
 
     public function __construct()
     {
@@ -30,11 +24,6 @@ class ComplexMathParser
         $this->parser = new Parser();
         $this->parser->setRationalFactory(true);
 
-    }
-
-    public function setSimplifying($flag)
-    {
-        $this->parser->setSimplifying($flag);
     }
 
     /**
@@ -50,22 +39,5 @@ class ComplexMathParser
 
         return $this->tree;
     }
-
-    /**
-     * Return the token list for the last parsed expression.
-     * @retval Token[]
-     */
-    public function getTokenList()
-    {
-        return $this->tokens;
-    }
-
-    /**
-     * Return the AST of the last parsed expression.
-     * @retval Node
-     */
-    public function getTree()
-    {
-        return $this->tree;
-    }
+    
 }
