+++ src/MathParser/StdMathParser.php
@@ -1,13 +1,9 @@
-<?php namespace MathParser;
+<?php 
+
+namespace MathParser;
 
 use MathParser\Lexing\StdMathLexer;
-use MathParser\Lexing\TokenDefinition;
-use MathParser\Lexing\TokenPrecedence;
-use MathParser\Lexing\TokenAssociativity;
-use MathParser\Lexing\TokenType;
 use MathParser\Parsing\Parser;
-use MathParser\Interpreting\Interpreter;
-use MathParser\Interpreting\PrettyPrinter;
 
 /**
  * Convenience class for using the MathParser library.
@@ -38,28 +34,15 @@ use MathParser\Interpreting\PrettyPrinter;
  * ~~~
  *
  */
-class StdMathParser
+class StdMathParser extends AbstractMathParser
 {
-    /** StdMathLexer $lexer instance of StdMathLexer used for tokenizing */
-    private $lexer;
-    /** Parser $parser instance of Parsed used for the actual parsing */
-    private $parser;
-    /** Token[] $tokens list of tokens generated by the Lexer */
-    private $tokens;
-    /** Node $tree AST generated by the parser */
-    private $tree;
 
     public function __construct()
     {
         $this->lexer = new StdMathLexer();
         $this->parser = new Parser();
     }
-
-    public function setSimplifying($flag)
-    {
-        $this->parser->setSimplifying($flag);
-    }
-
+    
     /**
      * Parse the given mathematical expression into an abstract syntax tree.
      *
@@ -73,22 +56,5 @@ class StdMathParser
 
         return $this->tree;
     }
-
-    /**
-     * Return the token list for the last parsed expression.
-     * @retval Token[]
-     */
-    public function getTokenList()
-    {
-        return $this->tokens;
-    }
-
-    /**
-     * Return the AST of the last parsed expression.
-     * @retval Node
-     */
-    public function getTree()
-    {
-        return $this->tree;
-    }
+    
 }
